Significantly sped up ground failure model runs by using gdal to resample

+++ tests/gfail_test.py
@@ -76,11 +76,12 @@ def test_zhu2015(tmpdir):
         shutil.copy(default_file+'_bak', default_file)
 
     # Remove backup and tempfile
-    os.remove(default_file+'_bak')
+    if os.path.exists(default_file+'_bak'):
+        os.remove(default_file+'_bak')
     shutil.rmtree(p)
 
     # Then do test
-    np.testing.assert_allclose(target_data, test_data)
+    np.testing.assert_allclose(target_data, test_data, rtol=1e-3)
 
 
 def test_zhu2015_web(tmpdir):
